Move the foresting configurations to a place that makes more sense.
Yetkin and I discussed whether the configurations should go in their own repository. However I think in order to make tagging a specific HiForest version as simple as possible there should only be one repository, and changes to the configuration should be tracked in that repo. This way you don't have to find out both the environment tag AND the configuration tag, since they will be the same.

Conflicts:
	ForestConfigurations/runForest_44X.py
	ForestConfigurations/runForest_44X_allTracks.py
	ForestConfigurations/runForest_44x_QM2012.py
	ForestConfigurations/runForest_MC_44X.py
	ForestConfigurations/runForest_MC_44X_allTracks.py
	ForestConfigurations/runForest_MC_onSIM_44X.py